Move indexing state enums to separate files
