+++ Kitodo/src/main/java/org/kitodo/production/enums/IndexStates.java
@@ -1,0 +1,20 @@
+/*
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ */
+
+package org.kitodo.production.enums;
+
+public enum IndexStates {
+    NO_STATE,
+    DELETE_ERROR,
+    DELETE_SUCCESS,
+    MAPPING_ERROR,
+    MAPPING_SUCCESS,
+}
